pyproject.toml
[tool.poetry]
name = "shop-bot"
version = "0.0.1"
description = "Bot for shop with webapp in telegram."
authors = ["Nikita Novikov <sqqvt1@gmail.com>"]
license = "GNU Affero General Public License v3.0"

[tool.poetry.dependencies]
python = "^3.10"
aiogram = "^2.23.1"
loguru = "^0.6.0"
python-dotenv = "^0.21.0"

[tool.poetry.dev-dependencies]
pytest = "^7.2.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
